Fix relative requirement bugs.
By default AtomLinter lints files in a temporary directory, for Crystal file relative requirement this is bad. So now you can run linter-crystal in a mode where it lints in your files actual directory.

The trade off here though is that the file will only be linted when it is first opened and when you save.

Both modes have their advantages, it's up to you to decide which works best for you.